Move schema names conversion code around
